% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_length_weight.R
\name{get_length_weight}
\alias{get_length_weight}
\title{Extract Length and weight data from svdbs}
\usage{
get_length_weight(channel, year = 1994, species = "all", sex = "all")
}
\arguments{
\item{channel}{an RODBC object (see \code{\link{connect_to_database}})}

\item{species}{a specific species code or set of codes. Either numeric or character vector. Defaults to "all" species.
Numeric codes are converted to VARCHAR2(3 BYTE) when creating the sql statement. Character codes are short character strings.}

\item{sex}{character vector. Default = "all". options "M" (male), "F" (female), "U" (unsexed)}
}
\value{
A list is returned:

  \item{data}{containing the result of the executed \code{$sql} statement}

  \item{sql}{containing the sql call}

  \item{colNames}{a vector of the table's column names}
}
\description{
Extract a list of length and weights for speices sampled in nefsc surveys (spring and fall)
This data is extracted from svdbs.union_fscs_svbio
}
\section{Reference}{

Use the data dictionary (\url{http://nova.nefsc.noaa.gov/datadict/}) for field name explanations.
Note: species codes (svspp) are stored in the database as VARCHAR2(3 BYTE)
}

\examples{
\dontrun{
# extracts info for cod (73)
channel <- connect_to_database(server="name_of_server",uid="individuals_username")
get_length_age(channel,species=73)

# extracts info for cod ("COD")
channel <- connect_to_database(server="name_of_server",uid="individuals_username")
get_length_age(channel,"cod") or
get_length_age(channel,"co") or
get_length_age(channel,"COD")

}

}
\seealso{
\code{\link{connect_to_database}}
}
