% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svdbs.R
\docType{package}
\name{svdbs}
\alias{svdbs}
\alias{svdbs-package}
\title{svdbs: Extract info from Database}
\description{
Tools to aid in extraction of data from the svdbs database
}
\details{
\itemize{
\item Get_ functions to access, species codes, survey indices, length-weight etc.
}

To learn more about using \code{svdbs}, start with the vignette: \code{browseVignettes(package="svdbs")} or click the index below
}
